Move user fixtures to the UserFixtures class
